+++ litellm/llms/OpenAI/openai.py
@@ -34,37 +34,8 @@ from litellm.utils import (
 
 from ...types.llms.openai import *
 from ..base import BaseLLM
-from ..prompt_templates.common_utils import convert_content_list_to_str
 from ..prompt_templates.factory import custom_prompt, prompt_factory
-from .common_utils import drop_params_from_unprocessable_entity_error
-from .completion.utils import is_tokens_or_list_of_tokens
-
-
-class OpenAIError(Exception):
-    def __init__(
-        self,
-        status_code,
-        message,
-        request: Optional[httpx.Request] = None,
-        response: Optional[httpx.Response] = None,
-        headers: Optional[httpx.Headers] = None,
-    ):
-        self.status_code = status_code
-        self.message = message
-        self.headers = headers
-        if request:
-            self.request = request
-        else:
-            self.request = httpx.Request(method="POST", url="https://api.openai.com/v1")
-        if response:
-            self.response = response
-        else:
-            self.response = httpx.Response(
-                status_code=status_code, request=self.request
-            )
-        super().__init__(
-            self.message
-        )  # Call the base class constructor with the parameters it needs
+from .common_utils import OpenAIError, drop_params_from_unprocessable_entity_error
 
 
 class MistralEmbeddingConfig:
@@ -379,155 +350,6 @@ class OpenAIConfig:
         )
 
 
-class OpenAITextCompletionConfig:
-    """
-    Reference: https://platform.openai.com/docs/api-reference/completions/create
-
-    The class `OpenAITextCompletionConfig` provides configuration for the OpenAI's text completion API interface. Below are the parameters:
-
-    - `best_of` (integer or null): This optional parameter generates server-side completions and returns the one with the highest log probability per token.
-
-    - `echo` (boolean or null): This optional parameter will echo back the prompt in addition to the completion.
-
-    - `frequency_penalty` (number or null): Defaults to 0. It is a numbers from -2.0 to 2.0, where positive values decrease the model's likelihood to repeat the same line.
-
-    - `logit_bias` (map): This optional parameter modifies the likelihood of specified tokens appearing in the completion.
-
-    - `logprobs` (integer or null): This optional parameter includes the log probabilities on the most likely tokens as well as the chosen tokens.
-
-    - `max_tokens` (integer or null): This optional parameter sets the maximum number of tokens to generate in the completion.
-
-    - `n` (integer or null): This optional parameter sets how many completions to generate for each prompt.
-
-    - `presence_penalty` (number or null): Defaults to 0 and can be between -2.0 and 2.0. Positive values increase the model's likelihood to talk about new topics.
-
-    - `stop` (string / array / null): Specifies up to 4 sequences where the API will stop generating further tokens.
-
-    - `suffix` (string or null): Defines the suffix that comes after a completion of inserted text.
-
-    - `temperature` (number or null): This optional parameter defines the sampling temperature to use.
-
-    - `top_p` (number or null): An alternative to sampling with temperature, used for nucleus sampling.
-    """
-
-    best_of: Optional[int] = None
-    echo: Optional[bool] = None
-    frequency_penalty: Optional[int] = None
-    logit_bias: Optional[dict] = None
-    logprobs: Optional[int] = None
-    max_tokens: Optional[int] = None
-    n: Optional[int] = None
-    presence_penalty: Optional[int] = None
-    stop: Optional[Union[str, list]] = None
-    suffix: Optional[str] = None
-    temperature: Optional[float] = None
-    top_p: Optional[float] = None
-
-    def __init__(
-        self,
-        best_of: Optional[int] = None,
-        echo: Optional[bool] = None,
-        frequency_penalty: Optional[int] = None,
-        logit_bias: Optional[dict] = None,
-        logprobs: Optional[int] = None,
-        max_tokens: Optional[int] = None,
-        n: Optional[int] = None,
-        presence_penalty: Optional[int] = None,
-        stop: Optional[Union[str, list]] = None,
-        suffix: Optional[str] = None,
-        temperature: Optional[float] = None,
-        top_p: Optional[float] = None,
-    ) -> None:
-        locals_ = locals().copy()
-        for key, value in locals_.items():
-            if key != "self" and value is not None:
-                setattr(self.__class__, key, value)
-
-    @classmethod
-    def get_config(cls):
-        return {
-            k: v
-            for k, v in cls.__dict__.items()
-            if not k.startswith("__")
-            and not isinstance(
-                v,
-                (
-                    types.FunctionType,
-                    types.BuiltinFunctionType,
-                    classmethod,
-                    staticmethod,
-                ),
-            )
-            and v is not None
-        }
-
-    def _transform_prompt(
-        self,
-        messages: Union[List[AllMessageValues], List[OpenAITextCompletionUserMessage]],
-    ) -> AllPromptValues:
-        if len(messages) == 1:  # base case
-            message_content = messages[0].get("content")
-            if (
-                message_content
-                and isinstance(message_content, list)
-                and is_tokens_or_list_of_tokens(message_content)
-            ):
-                openai_prompt: AllPromptValues = cast(AllPromptValues, message_content)
-            else:
-                openai_prompt = ""
-                content = convert_content_list_to_str(
-                    cast(AllMessageValues, messages[0])
-                )
-                openai_prompt += content
-        else:
-            prompt_str_list: List[str] = []
-            for m in messages:
-                try:  # expect list of int/list of list of int to be a 1 message array only.
-                    content = convert_content_list_to_str(cast(AllMessageValues, m))
-                    prompt_str_list.append(content)
-                except Exception as e:
-                    raise e
-            openai_prompt = prompt_str_list
-        return openai_prompt
-
-    def convert_to_chat_model_response_object(
-        self,
-        response_object: Optional[TextCompletionResponse] = None,
-        model_response_object: Optional[ModelResponse] = None,
-    ):
-        try:
-            ## RESPONSE OBJECT
-            if response_object is None or model_response_object is None:
-                raise ValueError("Error in response object format")
-            choice_list = []
-            for idx, choice in enumerate(response_object["choices"]):
-                message = Message(
-                    content=choice["text"],
-                    role="assistant",
-                )
-                choice = Choices(
-                    finish_reason=choice["finish_reason"], index=idx, message=message
-                )
-                choice_list.append(choice)
-            model_response_object.choices = choice_list
-
-            if "usage" in response_object:
-                setattr(model_response_object, "usage", response_object["usage"])
-
-            if "id" in response_object:
-                model_response_object.id = response_object["id"]
-
-            if "model" in response_object:
-                model_response_object.model = response_object["model"]
-
-            model_response_object._hidden_params["original_response"] = (
-                response_object  # track original response, if users make a litellm.text_completion() request, we can return the original response
-            )
-            return model_response_object
-        except Exception as e:
-            raise e
-
-
 class OpenAIChatCompletion(BaseLLM):
 
     def __init__(self) -> None:
@@ -710,7 +532,7 @@ class OpenAIChatCompletion(BaseLLM):
                         custom_llm_provider=custom_llm_provider,
                     )
             if messages is not None and custom_llm_provider is not None:
-                provider_config = ProviderConfigManager.get_provider_config(
+                provider_config = ProviderConfigManager.get_provider_chat_config(
                     model=model, provider=LlmProviders(custom_llm_provider)
                 )
                 messages = provider_config._transform_messages(messages)
@@ -1582,306 +1404,6 @@ class OpenAIChatCompletion(BaseLLM):
                 "x-ratelimit-remaining-tokens"
             ]
         return response
-
-
-class OpenAITextCompletion(BaseLLM):
-    openai_text_completion_global_config = OpenAITextCompletionConfig()
-
-    def __init__(self) -> None:
-        super().__init__()
-
-    def validate_environment(self, api_key):
-        headers = {
-            "content-type": "application/json",
-        }
-        if api_key:
-            headers["Authorization"] = f"Bearer {api_key}"
-        return headers
-
-    def completion(
-        self,
-        model_response: ModelResponse,
-        api_key: str,
-        model: str,
-        messages: Union[List[AllMessageValues], List[OpenAITextCompletionUserMessage]],
-        timeout: float,
-        logging_obj: LiteLLMLoggingObj,
-        optional_params: dict,
-        print_verbose: Optional[Callable] = None,
-        api_base: Optional[str] = None,
-        acompletion: bool = False,
-        litellm_params=None,
-        logger_fn=None,
-        client=None,
-        organization: Optional[str] = None,
-        headers: Optional[dict] = None,
-    ):
-        try:
-            if headers is None:
-                headers = self.validate_environment(api_key=api_key)
-            if model is None or messages is None:
-                raise OpenAIError(status_code=422, message="Missing model or messages")
-
-            # don't send max retries to the api, if set
-
-            prompt = self.openai_text_completion_global_config._transform_prompt(
-                messages
-            )
-
-            data = {"model": model, "prompt": prompt, **optional_params}
-            max_retries = data.pop("max_retries", 2)
-            ## LOGGING
-            logging_obj.pre_call(
-                input=messages,
-                api_key=api_key,
-                additional_args={
-                    "headers": headers,
-                    "api_base": api_base,
-                    "complete_input_dict": data,
-                },
-            )
-            if acompletion is True:
-                if optional_params.get("stream", False):
-                    return self.async_streaming(
-                        logging_obj=logging_obj,
-                        api_base=api_base,
-                        api_key=api_key,
-                        data=data,
-                        headers=headers,
-                        model_response=model_response,
-                        model=model,
-                        timeout=timeout,
-                        max_retries=max_retries,
-                        client=client,
-                        organization=organization,
-                    )
-                else:
-                    return self.acompletion(api_base=api_base, data=data, headers=headers, model_response=model_response, prompt=prompt, api_key=api_key, logging_obj=logging_obj, model=model, timeout=timeout, max_retries=max_retries, organization=organization, client=client)  # type: ignore
-            elif optional_params.get("stream", False):
-                return self.streaming(
-                    logging_obj=logging_obj,
-                    api_base=api_base,
-                    api_key=api_key,
-                    data=data,
-                    headers=headers,
-                    model_response=model_response,
-                    model=model,
-                    timeout=timeout,
-                    max_retries=max_retries,  # type: ignore
-                    client=client,
-                    organization=organization,
-                )
-            else:
-                if client is None:
-                    openai_client = OpenAI(
-                        api_key=api_key,
-                        base_url=api_base,
-                        http_client=litellm.client_session,
-                        timeout=timeout,
-                        max_retries=max_retries,  # type: ignore
-                        organization=organization,
-                    )
-                else:
-                    openai_client = client
-
-                raw_response = openai_client.completions.with_raw_response.create(**data)  # type: ignore
-                response = raw_response.parse()
-                response_json = response.model_dump()
-
-                ## LOGGING
-                logging_obj.post_call(
-                    input=prompt,
-                    api_key=api_key,
-                    original_response=response_json,
-                    additional_args={
-                        "headers": headers,
-                        "api_base": api_base,
-                    },
-                )
-
-                ## RESPONSE OBJECT
-                return TextCompletionResponse(**response_json)
-        except Exception as e:
-            status_code = getattr(e, "status_code", 500)
-            error_headers = getattr(e, "headers", None)
-            error_text = getattr(e, "text", str(e))
-            error_response = getattr(e, "response", None)
-            if error_headers is None and error_response:
-                error_headers = getattr(error_response, "headers", None)
-            raise OpenAIError(
-                status_code=status_code, message=error_text, headers=error_headers
-            )
-
-    async def acompletion(
-        self,
-        logging_obj,
-        api_base: str,
-        data: dict,
-        headers: dict,
-        model_response: ModelResponse,
-        prompt: str,
-        api_key: str,
-        model: str,
-        timeout: float,
-        max_retries: int,
-        organization: Optional[str] = None,
-        client=None,
-    ):
-        try:
-            if client is None:
-                openai_aclient = AsyncOpenAI(
-                    api_key=api_key,
-                    base_url=api_base,
-                    http_client=litellm.aclient_session,
-                    timeout=timeout,
-                    max_retries=max_retries,
-                    organization=organization,
-                )
-            else:
-                openai_aclient = client
-
-            raw_response = await openai_aclient.completions.with_raw_response.create(
-                **data
-            )
-            response = raw_response.parse()
-            response_json = response.model_dump()
-
-            ## LOGGING
-            logging_obj.post_call(
-                input=prompt,
-                api_key=api_key,
-                original_response=response,
-                additional_args={
-                    "headers": headers,
-                    "api_base": api_base,
-                },
-            )
-            ## RESPONSE OBJECT
-            response_obj = TextCompletionResponse(**response_json)
-            response_obj._hidden_params.original_response = json.dumps(response_json)
-            return response_obj
-        except Exception as e:
-            status_code = getattr(e, "status_code", 500)
-            error_headers = getattr(e, "headers", None)
-            error_text = getattr(e, "text", str(e))
-            error_response = getattr(e, "response", None)
-            if error_headers is None and error_response:
-                error_headers = getattr(error_response, "headers", None)
-            raise OpenAIError(
-                status_code=status_code, message=error_text, headers=error_headers
-            )
-
-    def streaming(
-        self,
-        logging_obj,
-        api_key: str,
-        data: dict,
-        headers: dict,
-        model_response: ModelResponse,
-        model: str,
-        timeout: float,
-        api_base: Optional[str] = None,
-        max_retries=None,
-        client=None,
-        organization=None,
-    ):
-
-        if client is None:
-            openai_client = OpenAI(
-                api_key=api_key,
-                base_url=api_base,
-                http_client=litellm.client_session,
-                timeout=timeout,
-                max_retries=max_retries,  # type: ignore
-                organization=organization,
-            )
-        else:
-            openai_client = client
-
-        try:
-            raw_response = openai_client.completions.with_raw_response.create(**data)
-            response = raw_response.parse()
-        except Exception as e:
-            status_code = getattr(e, "status_code", 500)
-            error_headers = getattr(e, "headers", None)
-            error_text = getattr(e, "text", str(e))
-            error_response = getattr(e, "response", None)
-            if error_headers is None and error_response:
-                error_headers = getattr(error_response, "headers", None)
-            raise OpenAIError(
-                status_code=status_code, message=error_text, headers=error_headers
-            )
-        streamwrapper = CustomStreamWrapper(
-            completion_stream=response,
-            model=model,
-            custom_llm_provider="text-completion-openai",
-            logging_obj=logging_obj,
-            stream_options=data.get("stream_options", None),
-        )
-
-        try:
-            for chunk in streamwrapper:
-                yield chunk
-        except Exception as e:
-            status_code = getattr(e, "status_code", 500)
-            error_headers = getattr(e, "headers", None)
-            error_text = getattr(e, "text", str(e))
-            error_response = getattr(e, "response", None)
-            if error_headers is None and error_response:
-                error_headers = getattr(error_response, "headers", None)
-            raise OpenAIError(
-                status_code=status_code, message=error_text, headers=error_headers
-            )
-
-    async def async_streaming(
-        self,
-        logging_obj,
-        api_key: str,
-        data: dict,
-        headers: dict,
-        model_response: ModelResponse,
-        model: str,
-        timeout: float,
-        max_retries: int,
-        api_base: Optional[str] = None,
-        client=None,
-        organization=None,
-    ):
-        if client is None:
-            openai_client = AsyncOpenAI(
-                api_key=api_key,
-                base_url=api_base,
-                http_client=litellm.aclient_session,
-                timeout=timeout,
-                max_retries=max_retries,
-                organization=organization,
-            )
-        else:
-            openai_client = client
-
-        raw_response = await openai_client.completions.with_raw_response.create(**data)
-        response = raw_response.parse()
-        streamwrapper = CustomStreamWrapper(
-            completion_stream=response,
-            model=model,
-            custom_llm_provider="text-completion-openai",
-            logging_obj=logging_obj,
-            stream_options=data.get("stream_options", None),
-        )
-
-        try:
-            async for transformed_chunk in streamwrapper:
-                yield transformed_chunk
-        except Exception as e:
-            status_code = getattr(e, "status_code", 500)
-            error_headers = getattr(e, "headers", None)
-            error_text = getattr(e, "text", str(e))
-            error_response = getattr(e, "response", None)
-            if error_headers is None and error_response:
-                error_headers = getattr(error_response, "headers", None)
-            raise OpenAIError(
-                status_code=status_code, message=error_text, headers=error_headers
-            )
 
 
 class OpenAIFilesAPI(BaseLLM):
